implement stack arrays
Currently they cannot be used in state{..} because it's initialized by a
separate function and the array does not persist.

A possible solution is to mock the stack in init() recursion using the
storage to extend all stack allocations lifetime indefinetly from step()
POV. Unfortunately this means *all* init() stack temporaries will stay in
storage. Not sur if it's ok.